Change Header.nBits to u32
nBits is a 32-bit type, so decode_compact_bits doesn't need a u64 argument. In Scala/ErgoTree it's i64 because Scala doesn't have an unsigned 32-bit type

+++ bindings/ergo-lib-python/src/chain/header.rs
@@ -81,7 +81,7 @@ impl Header {
         self.0.timestamp
     }
     #[getter]
-    fn n_bits(&self) -> u64 {
+    fn n_bits(&self) -> u32 {
         self.0.n_bits
     }
     #[getter]
